% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNASeq_datasets.R
\docType{data}
\name{miRNASeq.20160128}
\alias{BLCA.miRNASeq.20160128}
\alias{BLCA.miRNASeq.20160128,BRCA.miRNASeq.20160128,COADREAD.miRNASeq.20160128,HNSC.miRNASeq.20160128,KIPAN.miRNASeq.20160128,KIRC.miRNASeq.20160128,KIRP.miRNASeq.20160128,LAML.miRNASeq.20160128,LUAD.miRNASeq.20160128,LUSC.miRNASeq.20160128,STAD.miRNASeq.20160128,STES.miRNASeq.20160128,UCEC.miRNASeq.20160128}
\alias{BRCA.miRNASeq.20160128}
\alias{COADREAD.miRNASeq.20160128}
\alias{HNSC.miRNASeq.20160128}
\alias{KIPAN.miRNASeq.20160128}
\alias{KIRC.miRNASeq.20160128}
\alias{KIRP.miRNASeq.20160128}
\alias{LAML.miRNASeq.20160128}
\alias{LUAD.miRNASeq.20160128}
\alias{LUSC.miRNASeq.20160128}
\alias{STAD.miRNASeq.20160128}
\alias{STES.miRNASeq.20160128}
\alias{UCEC.miRNASeq.20160128}
\alias{miRNASeq.20160128}
\title{miRNASeq datasets from TCGA project from 2016-01-28 release date}
\source{
\url{http://gdac.broadinstitute.org/}
}
\usage{
BLCA.miRNASeq.20160128

BRCA.miRNASeq.20160128

COADREAD.miRNASeq.20160128

HNSC.miRNASeq.20160128

KIPAN.miRNASeq.20160128

KIRC.miRNASeq.20160128

KIRP.miRNASeq.20160128

LAML.miRNASeq.20160128

LUAD.miRNASeq.20160128

LUSC.miRNASeq.20160128

STAD.miRNASeq.20160128

STES.miRNASeq.20160128

UCEC.miRNASeq.20160128
}
\arguments{
\item{metadata}{A logical indicating whether load data into the workspace (default, \code{FALSE}) or to only display the object's metadata (\code{TRUE}). See examples.}
}
\description{
Package provides miRNASeq datasets from The Cancer Genome Atlas Project for all cohorts types from \url{http://gdac.broadinstitute.org/}.
Data were downloaded using \link{RTCGA-package} and contain snapshots for the release date: \code{2016-01-28} . Visit \pkg{RTCGA} site: \url{http://rtcga.github.io/RTCGA/}.
Use cases, examples and information about datasets in \pkg{RTCGA.miRNASeq.20160128}  can be found here: \code{browseVignettes("RTCGA")}. Link to the data format explanation is in the package DESCRIPTION file.
}
\details{
\code{browseVignettes("RTCGA")}
}
\examples{

\dontrun{
 BLCA.miRNASeq.20160128(metadata = TRUE) 
 BLCA.miRNASeq.20160128(metadata = FALSE) 
 BLCA.miRNASeq.20160128 
}

}
\keyword{datasets}

